Move Android review reactions into package
