+++ apps/android/feature/review/src/main/java/com/flashcardsopensourceapp/feature/review/ReviewScreenTags.kt
@@ -15,4 +15,3 @@ const val reviewCurrentCardTag: String = "review_current_card"
 const val reviewCurrentCardFrontContentTag: String = "review_current_card_front_content"
 const val reviewProgressBadgeTag: String = "review_progress_badge"
 const val reviewQueueButtonTag: String = "review_queue_button"
-const val testAnimationsScreenTag: String = "test_animations_screen"
